Fix broken reverse relationship on models with history
We found that adding history to a model with a foreign key resulted
in the reverse relationship disappearing (ie modelname_set no longer
existed).

The fix is to make a deeper copy of the field, so that when:

    field.rel.related_name = '+'

is run, this only changes the local copy, not the related name of the
field in the original model.