Add tests for setting verbose name
The motivation for this is that the auth_permission table has a maximum
length of 50 - see https://code.djangoproject.com/ticket/17763

We had a class with a long name - prepending "Historical" to the class
name meant that the name of the permission was over 50 characters long.

Once these tests pass, we will be able to set a shorter verbose_name and
hence workaround the above bug.