Change string types to work with both python 2 and 3
Because we do:

    from future import unicode_literals

The string 'historical ' will be unicode.  So when it is concatanated
with the following str(..) the combined string will still be unicode.

But we still need to use str(..) as model._meta.verbose_name is not a
string to start with, but a proxy object.

In Python 3 it is just all unicode, the but the unicode() builtin is not
available.