Ensure custom arguments for fields are included in historical models' fields.
Previously, if a custom field class was used for a foreign key and that
class accepted custom kwargs, those kwargs would not be included in the
creation of the historical model's field and would break the migration.
To address this, we'll use the deconstruct() method of the field to get
the arguments that should be passed to it and override the ones we need
to for the historical model's field.