Added tracked_fields attr to historical models
This attribute can be useful for future code and for users, as there is currently no easy way (known to me) to list
the fields of a model that are tracked (i.e. not excluded) by a history model.

The attribute will also be used in an upcoming bugfix, which was the main motivation for adding this.