Add a history_repr field that stores the value of __str__()
Sometimes __str__() requires of a foreign key to represent its value.
In those scenarios, two issues appear:

1. Need to perform extra queries. select_related() does not help since
the relationship exists in the model, not the historical model.
2. If the related instance is removed, it will fail with DoesNotExist.

One way of overcoming this situation is by storing the value of __str__()
at the time we create the Historical Record.

Since this might have a performance penalty, it MUST be explicitly
enabled by setting `SIMPLE_HISTORY_HISTORIC_REPR` to True.