Handle if `request` doesn't exist on the context in middleware
The change in https://github.com/jazzband/django-simple-history/pull/1188
removed a check for if the request context had a "request" entry in it before
trying to delete the request data.

This has a side effect of breaking tests in order projects in very specific
circumstances, where multiple requests pass through the middleware,
concurrently, inside the same thread, as it means the middleware then may try
to remove the request key on a context where it already has been removed.

This shouldn't have any impact in real life scenarios.