Attempts to fix update_change_reason for nullable JSONField fields
As reported in issue #1181, the implementation of the utility `update_change_reason()`
method does not work for models with a nullable JSONField. This is caused by the
slightly more convoluted querying mechanisms that have to be applied for JSONFields,
please cf. here:
[django dev documentation](https://docs.djangoproject.com/en/dev/topics/db/queries/#querying-jsonfield)

This attempts to fix this by assuming that
- a nullable JSONField with a `value` of `None` is not set at all
- a non-nullable JSONField with a `value` of `None` is actually a
    "NULL" dictionary

I am aware that this is not the only way to handle this. But it is the convention I am
using in my project and this fixed the issue for me. It may be worthwhile to implement a
policy switch to decide how to interpret `None` in this specific case.