Made object history list template overridable
Including renaming it to remove the `_` prefix (indicating that it's
internal / not meant for usage by library users), and adding an
`object_history_list_template` field to the admin class.
This rename also changed the order of messages in the translation files.

Also deprecated the template tag `simple_history_admin_list.display_list`,
as it's now entirely unused within the project.