Refactored formatting history delta changes
It's now easier to only override formatting `ModelChange.old` and
`ModelChange.new`, and do so before the values are converted into
strings and truncated - which was previously (almost) only possible if
the user discarded our implementation of
`format_history_delta_change()`.

Also added documentation on ways of customizing the admin class'
template context, in `admin.rst`.