Added Changes column to admin history page
This lists the changes between each historical record of the object.

Effort has been made to make it easy for library users to override most
parts of this, and to let them customize only the details they want to
without having to copy or replace our implementation.
An overview of how to override it has been added to `admin.rst`.

Most of the code was added to `template_utils.py`, to avoid cluttering
`SimpleHistoryAdmin`.

Also updated the Norwegian Bokmål translations,
and updated all the docs screenshots, to make them have Django 5.0's UI
and to showcase the new column.

Co-authored-by: Gur Raunaq Singh <raunaqsoni.dev123@gmail.com>
Co-authored-by: Tim Schilling <schilling711@gmail.com>