Deprecated HistoricalRecords.thread
`HistoricalRecords.context` was added in favor of it a few years ago,
and since it has always been undocumented, it's time to properly
deprecate it.
4.0 would have been the natural version to remove it if it were
documented as part of the publicly exposed API, but since it's not,
3.10 seems like a reasonable choice.

Also removed importing `threading.local` in case `asgiref.local.Local`
is not available, as the latter should always be installed now that all
our supported Django versions use it as a requirement.